Struct coset::CoseMac0

pub struct CoseMac0 {
    pub protected: ProtectedHeader,
    pub unprotected: Header,
    pub payload: Option<Vec<u8>>,
    pub tag: Vec<u8>,
}

Structure representing a message with authentication code (MAC) where the relevant key is implicit.

 COSE_Mac0 = [
    Headers,
    payload : bstr / nil,
    tag : bstr,
 ]

Fields
protected: ProtectedHeaderunprotected: Headerpayload: Option<Vec<u8>>tag: Vec<u8>
Implementations

impl CoseMac0

pub fn verify_tag<F, E>(&self, external_aad: &[u8], verify: F) -> Result<(), E>where F: FnOnce(&[u8], &[u8]) -> Result<(), E>,

Verify the tag value using the provided mac function, feeding it the tag value and the combined to-be-MACed data (in that order).

Panics

This function will panic if the payload has not been set.