Struct coset::HeaderBuilder

pub struct HeaderBuilder(_);

Builder for Header objects.

Implementations

impl HeaderBuilder

pub fn new() -> Self

Constructor for builder.

pub fn build(self) -> Header

Build the completed object.

pub fn key_id(self, key_id: Vec<u8>) -> Self

Set the associated field.

pub fn algorithm(self, alg: Algorithm) -> Self

Set the algorithm.

pub fn add_critical(self, param: HeaderParameter) -> Self

Add a critical header.

pub fn add_critical_label(self, label: RegisteredLabel<HeaderParameter>) -> Self

Add a critical header.

pub fn content_format(self, content_type: CoapContentFormat) -> Self

Set the content type to a numeric value.

pub fn content_type(self, content_type: String) -> Self

Set the content type to a text value.

pub fn iv(self, iv: Vec<u8>) -> Self

Set the IV, and clear any partial IV already set.

pub fn partial_iv(self, iv: Vec<u8>) -> Self

Set the partial IV, and clear any IV already set.

pub fn add_counter_signature(self, sig: CoseSignature) -> Self

Add a counter signature.

pub fn value(self, label: i64, value: Value) -> Self

Set a header label:value pair. If duplicate labels are added to a Header, subsequent attempts to CBOR-encode the header will fail.

Panics

This function will panic if it used to set a header label from the range [1, 6].

pub fn text_value(self, label: String, value: Value) -> Self

Set a header label:value pair where the label is text.