Trait coset::TaggedCborSerializable

pub trait TaggedCborSerializable: AsCborValue {
    const TAG: u64;
    fn from_tagged_slice(slice: &[u8]) -> Result<Self, CoseError> { ... }

    fn to_tagged_vec(self) -> Result<Vec<u8>, CoseError> { ... }
}

Extension trait that adds tagged serialization/deserialization methods.

Associated Constants

const TAG: u64

The associated tag value.

Provided methods

fn from_tagged_slice(slice: &[u8]) -> Result<Self, CoseError>

Create an object instance from serialized CBOR data in a slice, expecting an initial tag value.

fn to_tagged_vec(self) -> Result<Vec<u8>, CoseError>

Serialize this object to a vector, including initial tag, consuming the object along the way.