Module coset::iana

Enumerations for IANA-managed values.

Sources:

https://www.iana.org/assignments/cose/cose.xhtml

https://www.iana.org/assignments/cbor-tags/cbor-tags.xhtml

https://www.iana.org/assignments/core-parameters/core-parameters.xhtml#content-formats

https://www.iana.org/assignments/cwt/cwt.xhtml

Enums

Algorithm

IANA-registered COSE algorithms.

CborTag

CBOR tag values for COSE structures.

CoapContentFormat

CoAP Content Formats

CwtClaimName

CBOR Web Token (CWT) Claims From IANA registry https://www.iana.org/assignments/cwt/cwt.xhtml as of 2021-10-21.

Ec2KeyParameter

IANA-registered COSE key parameters for keys of type KeyType::EC2.

EllipticCurve

IANA-registered COSE elliptic curves.

HeaderAlgorithmParameter

IANA-registered COSE header algorithm parameters.

HeaderParameter

IANA-registered COSE header parameters.

HssLmsKeyParameter

IANA-registered COSE key parameters for keys of type KeyType::HSS_LMS.

KeyOperation

Key operation values.

KeyParameter

IANA-registered COSE common key parameters.

KeyType

IANA-registered COSE key types.

OkpKeyParameter

IANA-registered COSE key parameters for keys of type KeyType::OKP.

RsaKeyParameter

IANA-registered COSE key parameters for keys of type KeyType::RSA.

SymmetricKeyParameter

IANA-registered COSE key parameters for keys of type KeyType::Symmetric.

WalnutDsaKeyParameter

IANA-registered COSE key parameters for keys of type KeyType::WalnutDSA.

Constants

ALGORITHM_PRIVATE_USE_MAX

Integer values for COSE algorithms below this value are reserved for private use.

CWT_CLAIM_PRIVATE_USE_MAX

Integer values for CWT claims below this value are reserved for private use.

ELLIPTIC_CURVE_PRIVATE_USE_MAX

Integer values for COSE elliptic curves below this value are reserved for private use.

HEADER_PARAMETER_PRIVATE_USE_MAX

Integer values for COSE header parameters below this value are reserved for private use.

Traits

EnumI64

Trait indicating an enum that can be constructed from i64 values.

WithPrivateRange

Trait indicating an enum with a range of private values.