Struct coset::CoseKeyBuilder

pub struct CoseKeyBuilder(_);

Builder for CoseKey objects.

Implementations

impl CoseKeyBuilder

pub fn new() -> Self

Constructor for builder.

pub fn build(self) -> CoseKey

Build the completed object.

pub fn key_id(self, key_id: Vec<u8>) -> Self

Set the associated field.

pub fn base_iv(self, base_iv: Vec<u8>) -> Self

Set the associated field.

pub fn new_ec2_pub_key(curve: EllipticCurve, x: Vec<u8>, y: Vec<u8>) -> Self

Constructor for an elliptic curve public key specified by x and y coordinates.

pub fn new_ec2_pub_key_y_sign(
curve: EllipticCurve,
x: Vec<u8>,
y_sign: bool
) -> Self

Constructor for an elliptic curve public key specified by x coordinate plus sign of y coordinate.

pub fn new_ec2_priv_key(
curve: EllipticCurve,
x: Vec<u8>,
y: Vec<u8>,
d: Vec<u8>
) -> Self

Constructor for an elliptic curve private key specified by d, together with public x and y coordinates.

pub fn new_symmetric_key(k: Vec<u8>) -> Self

Constructor for a symmetric key specified by k.

pub fn algorithm(self, alg: Algorithm) -> Self

Set the algorithm.

pub fn add_key_op(self, op: KeyOperation) -> Self

Add a key operation.

pub fn param(self, label: i128, value: Value) -> Self

Set a parameter value.

Panics

This function will panic if it used to set a parameter label from the iana::KeyParameter range.